% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateReads.R
\name{annotateReads}
\alias{annotateReads}
\title{Annotate reads as REF or ALT}
\usage{
annotateReads(ga, targets = NULL, refGenome = "hg38")
}
\arguments{
\item{ga}{list containing a GenomicAlignments object (see getReadsFromBAM)}

\item{targets}{data.frame containing target mutations (see formatTargets)}

\item{refGenome}{BSgenome to use (required for evaluating insertions; currently supports hg19 and hg38)}
}
\value{
list containing original and annotated GenomicAlignments objects
}
\description{
This function annotates each read in a given GenomicAlignment object as having the expected REF
or ALT allele (as provided in targets file).
}
\examples{
test.maf.file <- system.file('extdata', 'test_snp_data.maf', package = 'CompareFragmentSize'); 
test.bam.file <- system.file('extdata', 'sample1_example.bam', package = 'CompareFragmentSize'); 
mutations <- formatTargets(input = test.maf.file, id = 'Sample1');
target.reads <- getReadsFromBAM(filePath = test.bam.file, targets = mutations[1,]);
fragment.data <- annotateReads(ga = target.reads, targets = mutations[1,], refGenome = 'hg19');
}
