% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectMetrics.R
\name{collectMetrics}
\alias{collectMetrics}
\title{Summarize fragment size}
\usage{
collectMetrics(fragment.data, target, verbose = TRUE)
}
\arguments{
\item{fragment.data}{list containing annotated fragment data generated by annotateReads}

\item{target}{data.frame containing target mutation (see formatTargets)}

\item{verbose}{logical indicating whether or not to print metrics to screen (default = TRUE)}
}
\value{
data.frame of summary metrics
}
\description{
This function returns summary metrics, including counts and median fragment size, and attempts
to classify the target variant as somatic, germline or non-somatic.
}
\examples{
test.maf.file <- system.file('extdata', 'test_snp_data.maf', package = 'CompareFragmentSize'); 
test.bam.file <- system.file('extdata', 'sample1_example.bam', package = 'CompareFragmentSize'); 
mutations <- formatTargets(input = test.maf.file, id = 'Sample1');
target.reads <- getReadsFromBAM(filePath = test.bam.file, targets = mutations[1,]);
fragment.data <- annotateReads(ga = target.reads, targets = mutations[1,]);
summary.metrics <- collectMetrics(fragment.data, mutations[1,]);
}
