% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReadsFromBAM.R
\name{getReadsFromBAM}
\alias{getReadsFromBAM}
\title{Load BAM and extract reads}
\usage{
getReadsFromBAM(filePath, mapQthreshold = 30, targets = NULL, dedup = FALSE)
}
\arguments{
\item{filePath}{path to BAM file}

\item{mapQthreshold}{threshold to filter low quality reads (default = 30)}

\item{targets}{data.frame containing target mutations (see formatTargets)}

\item{dedup}{logical indicating whether to remove duplicates or not (default = FALSE)}
}
\value{
a genomic alignments object
}
\description{
This function loads a processed BAM file and subsets reads to thos overlapping target positions.
}
\examples{
test.maf.file <- system.file('extdata', 'test_snp_data.maf', package = 'CompareFragmentSize'); 
test.data.maf <- formatTargets(input = test.maf.file, id = 'Sample1');
test.bam.file <- system.file('extdata', 'sample1_example.bam', package = 'CompareFragmentSize'); 
target.reads <- getReadsFromBAM(filePath = test.bam.file, targets = test.data.maf[1,]);
}
