% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFragmentSizes.R
\name{plotFragmentSizes}
\alias{plotFragmentSizes}
\title{Plot cumulative fragment size distributions}
\usage{
plotFragmentSizes(fragment.data, target, id = NULL, output.to = "file")
}
\arguments{
\item{fragment.data}{list containing annotated fragment data generated by annotateReads}

\item{target}{data.frame containing target mutation (see formatTargets)}

\item{id}{sample ID; used for plot title (default = NULL)}

\item{output.to}{should the plot be output to file (as a png) or output to screen (default = 'file')}
}
\value{
png file or print image to screen
}
\description{
This function generates a cumulative distribution plot comparing fragment size between
REF- and ALT-containing reads using ggplot2.
}
\examples{
test.maf.file <- system.file('extdata', 'test_snp_data.maf', package = 'CompareFragmentSize'); 
test.bam.file <- system.file('extdata', 'sample1_example.bam', package = 'CompareFragmentSize'); 
mutations <- formatTargets(input = test.maf.file, id = 'Sample1');
target.reads <- getReadsFromBAM(filePath = test.bam.file, targets = mutations[1,]);
fragment.data <- annotateReads(ga = target.reads, targets = mutations[1,]);
summary.metrics <- collectMetrics(fragment.data, mutations[1,]);
#not run: plotFragmentSizes(fragment.data, 'Sample1', summary.metrics);
}
